🐛 Fix: 코드 merge 수정
