+++ src/components/common/tab/PresetTab/PresetTab.tsx
@@ -4,7 +4,6 @@ import { Tabs, TabsContent, TabsList } from '@/components/common/ui/tabs';
 import { Dispatch, SetStateAction, useState } from 'react';
 import { Category as PresetCategory } from '@/constants/Category';
 
-
 interface PresetItem {
   id: number;
   description: string;
@@ -22,8 +21,8 @@ interface PresetTabProps {
   handleSettingClick?: (itemId: number) => void;
   handleDeleteClick?: (itemId: number) => void;
   isBottomSheet?: boolean;
-  setSelectedHouseWork?: Dispatch<SetStateAction<string | null>>;
-  setSelectedCategory?: Dispatch<SetStateAction<string | null>>;
+  handleClick: (id: number, description: string, category: string) => void;
+  selectedItem: number | null;
 }
 
 const PresetTab: React.FC<PresetTabProps> = ({
@@ -33,21 +32,9 @@ const PresetTab: React.FC<PresetTabProps> = ({
   handleSettingClick,
   handleDeleteClick,
   isBottomSheet = false,
-  setSelectedHouseWork,
-  setSelectedCategory,
+  handleClick,
+  selectedItem,
 }) => {
-  const [selectedItem, setSelectedItem] = useState<number | null>(null);
-
-  const handleClick = (id: number, description: string, category: string) => {
-    setSelectedItem(selectedItem === id ? null : id); //id가 같으면 선택해제되도록
-    if (setSelectedHouseWork) {
-      setSelectedHouseWork(description);
-    }
-    if (setSelectedCategory) {
-      setSelectedCategory(category);
-    }
-  };
-
   const allPresetData = {
     category: PresetCategory.ALL,
     items: data.flatMap(category =>
@@ -73,11 +60,12 @@ const PresetTab: React.FC<PresetTabProps> = ({
             <PresetItem
               category={item.category}
               housework={item.description}
-              handleSelectClick={() => handleClick(item.description)}
+              handleSelectClick={() => handleClick(item.id, item.description, item.category)}
               isInPresetSetting={isInPresetSetting}
               isShowDeleteBtn={deleteButtonStates[item.id]} //각 아이템의 boolean값이 들어간다.
               handleSettingClick={handleSettingClick && (() => handleSettingClick(item.id))}
               handleDeleteClick={handleDeleteClick && (() => handleDeleteClick(item.id))}
+              isSelected={selectedItem === item.id}
             />
           </div>
         ))}
